First, changed migration column type of SSN from integer to bigInteger. Second added getAddress method to patient class and update edit method on patient controller. Third created new patterns curp and ssn imask on beagle.js. Finally created index, create, edit and show patient views.

+++ database/migrations/2019_09_16_165933_create_ssns_table.php
@@ -17,7 +17,7 @@ class CreateSsnsTable extends Migration
             $table->uuid('id')->primary();
             $table->unsignedInteger('ssn_type_id');
             $table->string('ssn');
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->string('kinship');
             $table->date('date_start')->nullable();
             $table->date('date_end')->nullable();
